Semi working version of a tiddlyweb api explorer.
This uses documentation on tiddlyweb.tiddlyspace.com as the
base from which to know routes.

You can choose a route, fill it in with some data and then make
a GET request. For other types, you can see that it is possible
DELETE, POST or PUT but not actually do it.

Rather rough around the edges so far.